Changing the signature of similarity coefficient calculations

+++ dclsuite/src/dclsuite/resolution/similarity/BaroniUrbaniCoefficientStrategy.java
@@ -8,7 +8,7 @@ public class BaroniUrbaniCoefficientStrategy implements ICoefficientStrategy {
 	}
 	
 	@Override
-	public double calculate(double p, double q, double r, double s) {
+	public double calculate(int p, int q, int r, int s) {
 		return (Math.sqrt(p * s) + p) / (Math.sqrt(p * s) + (p + q) + (p + r) - p);
 	}
