+++ dclsuite/src/dclsuite/resolution/similarity/MountfordCoefficientStrategy.java
@@ -8,8 +8,8 @@ public class MountfordCoefficientStrategy implements ICoefficientStrategy {
 	}
 	
 	@Override
-	public double calculate(double p, double q, double r, double s) {
-		return (2 * p) / ((2 * (p + q) * (p + r)) - (((p + q) + (p + r)) * p));
+	public double calculate(int p, int q, int r, int s) {
+		return (2.0 * p) / ((2 * (p + q) * (p + r)) - (((p + q) + (p + r)) * p));
 	}
 
 }
